+++ src/modules/integrations/process-proxy.nix
@@ -1,0 +1,8 @@
+{ lib, ... }:
+{
+  options.process.proxy.enable = lib.mkEnableOption ''
+    the shared HTTP proxy for friendly process URLs under ``.localhost``
+  '' // {
+    default = true;
+  };
+}